rename formation_names.h to battle_names.h
