+++ src/battle/battle_names.h
